Using createReactWrapper and passing props to react
