+++ app/javascript/packs/LoginPage.js
@@ -1,8 +1,11 @@
 import React from 'react'
 import { render } from 'react-dom'
 import '@patternfly/react-core/dist/styles/base.css'
-import SimpleLoginPage from 'LoginPage/index'
+import {LoginPageWrapper} from 'LoginPage'
+import {safeFromJsonString} from 'utilities/json-utils'
 
 document.addEventListener('DOMContentLoaded', () => {
-  render(<SimpleLoginPage />, document.getElementById('login-page-wrapper'))
+  const loginPageContainer = document.getElementById('login-page-wrapper')
+  const loginPageProps = safeFromJsonString(loginPageContainer.dataset.loginProps)
+  LoginPageWrapper(LoginPageWrapper, 'login-page-wrapper')
 })
