+++ app/views/provider/sessions/_request_password_reset_form.html.slim
@@ -17,6 +17,6 @@
         li.forgot
           = link_to provider_admin_dashboard_path do
             i.fa.fa-undo>
-            | Sign in
+            | Sign in 
         li.commit
           = submit_tag 'Reset password', class: 'important-button'
